+++ Tuneful/Views/MiniPlayer/MiniPlayerView.swift
@@ -62,11 +62,14 @@ struct MiniPlayerView: View {
                 )
                 .opacity(showPlayerWindow ? 1 : 0)
                 .onChange(of: showPlayerWindow) { show in
-                    if show {
+                    if !show {
                         playerManager.startTimer()
                     } else {
                         playerManager.stopTimer()
                     }
+                }
+                .onAppear {
+                    playerManager.startTimer()
                 }
             }
         }
