+++ Tuneful/Views/Settings/MiniPlayerSettingsView.swift
@@ -29,6 +29,9 @@ struct MiniPlayerSettingsView: View {
                         Defaults.Toggle("", key: .showPlayerWindow)
                             .toggleStyle(.switch)
                             .controlSize(.small)
+                            .onChange(of: showPlayerWindow) { _ in
+                                NSApplication.shared.sendAction(#selector(AppDelegate.togglePlayerMenuItem), to: nil, from: nil)
+                            }
                     }
                     .padding(8)
